Implementing catch experience: Introducing functions
Here we made a new function called ApplyExperienceAfterEnemyCaught. Basically, everything is copied except anything related to Trainers, since you wouldn't capture a Trainer's Pokémon ever (you don't want to be a thief), and also the check to see if you have a fit Pokémon (it's already a precondition to throw a ball in the first place). Moreover, there's no need to update the battle scene or change the music to the victory jingle since the game already does this for us once the wild Pokémon is successfully captured. Everything else, including calculating boosts and Exp. Share divisions, is done in the function GiveExperiencePoints.