Changed Johto Wild Encounters
Sentret replaces Rattata on most routes in Day (Furret replaces Raticate), with the rats keeping most caves and Night encounters.

Either Murkrow or Misdreavus become night encounters on most routes after Route 34 (Murkrow also in Ilex).

Marill is now an early route pokemon, is available more often outside of Mt. Mortar and is a more common Surf encounter.

Houndour is the night encounter for Growlithe/Vulpix on their respective routes, and also appears in the Burned Tower.

Slugma is an encounter in the basement of the Burned Tower as well as in Dark Cave.

Sneasel is available on the route outside Ice Path (very rare) as well as inside it. Delibird and Swinub appear at all times of day in different rarity.

Larvitar is a rare encounter in Mt. Mortar.

Remoraid is fishable on Route 44 where he was originally, but also Ocean routes (41, 26, 27 and New Bark) with a Super Rod

Corsola is a rare Rock Smash chance as well as a Fish chance.